+++ src/components/common/header/IntroHeaderComponent.tsx
@@ -1,13 +1,13 @@
 import * as React from "react";
 import { Link, navigate } from "gatsby";
 import { recentPostDataType } from "../../layout/recent/recentPostsData";
-import { initialFilterKeyState } from "./filterKey";
+import { initialFilterKeyState } from "./filter/filterKey";
 import { LOGO_IMG_URL, LOGO_TEXT } from "../../../constant/constant";
 import { useSearch, useScroll, useSearchParam } from "../../../hooks";
-import FilterBoxComponent from "./FilterBoxComponent";
-import DarkModeSwitchComponent from "./DarkModeSwitchComponent";
+import FilterBoxComponent from "./filter";
+import DarkModeSwitchComponent from "./darkMode";
 import { useRecoilState } from "recoil";
-import { sessionState } from "../../../recoil/session/session";
+import { sessionState } from "../../../recoil/session";
 import { useSmoothScroll } from "../../../hooks";
 
 // style
