+++ src/pages/index.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 
 //components
-import SEOComponent from "../components/common/seoComponent";
+import SEOComponent from "../components/common/seo";
 import HeaderComponent from "../components/common/header";
-import PostComment from "../components/layout/comment/comment";
+import PostComment from "../components/layout/comment";
 import PopularComponent from "../components/layout/popular";
 import RecentComponent from "../components/layout/recent";
